+++ src/test/MainTest.java
@@ -20,16 +20,16 @@ public class MainTest extends TestCase {
     @Before
     public void init() { // Modifiquem els comportaments als que nosaltres volem
         when(mM.setEntradaTec(1)).thenReturn("Plus");
-        when(mM.setEntradaValor1(1)).thenReturn(1.0f);
-        when(mM.setEntradaValor2(2)).thenReturn(2.0f);
-        when(mM.finalResult("Plus",1.0f,2.0f)).thenReturn(3.0f);
+        when(mM.setEntradaValor1(1)).thenReturn(1.0);
+        when(mM.setEntradaValor2(2)).thenReturn(2.0);
+        when(mM.finalResult("Plus",1.0f,2.0f)).thenReturn(3.0);
     }
     @Test
     public void mockEntradaTeclado() {
         String entradaTecladoOp = s.setEntradaTecladoOp(1); //entradaTecladoOp té "Suma"
-        float valor1 = s.setEntradaValor1(1); //valor1 té 1.0;
-        float valor2 = s.setEntradaValor2(2); //valor2 té 2.0;
-        float resultadoFinal = s.resultadoFinal(entradaTecladoOp, valor1, valor2); // resultadoFinal té 3.0
+        double valor1 = s.setEntradaValor1(1); //valor1 té 1.0;
+        double valor2 = s.setEntradaValor2(2); //valor2 té 2.0;
+        double resultadoFinal = s.resultadoFinal(entradaTecladoOp, valor1, valor2); // resultadoFinal té 3.0
         assertEquals("Plus",entradaTecladoOp);
         assertEquals(1.0,valor1);
         assertEquals(2.0,valor2);
